코드 갱신

+++ Scripts/Interface/IDamages.cs
@@ -2,5 +2,5 @@
 using UnityEngine;
 public interface IDamages
 {
-    public void Damage(int hitDmg);
+    public void Damage(float hitDmg);
 }
